fix compilation in Cygwin GCC
Cygwin GCC doesn't like extern declaration of optind and opterr, and
reports compilation errors. Removing them doesn't seem to have any
adverse effects.

strdup() is a GNU extension, and this too causes problems in Cygwin GCC.
Exporting _GNU_SOURCE when compiling in Cygwin makes it available. This
must be conditional, otherwise it'll cause errors regarding redefinition
of a symbol in Linux.

Tested in GCC 10.2 in Cygwin.