csvscrob: add option to enable UTC output
By default it still outputs the time in local timezone, but can be
switched to UTC if needed.

Feature is implemented by using gmtime() instead of localtime() when
epoch scrobble time is converted to human-readable form. The selection
of funcion is done with function pointers.

This feature is useful when privacy.resistFingerprinting is enabled in
Firefox.